+++ src/main/java/rat/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package rat.storage;
+package rat.tasks;
 
 /**
  * This class encapsulates a Deadline task.
