+++ src/main/java/rat/tasks/Event.java
@@ -1,4 +1,4 @@
-package rat.storage;
+package rat.tasks;
 
 /**
  * This class encapsulates an Event task.
