+++ src/main/java/rat/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package rat.storage;
+package rat.tasks;
 
 /**
  * This class encapsulates a ToDo task.
